u2fd: Fix char to int casting bug
Current code in u2fhid casts a char into a signed int, but in fact that
byte should be interpreted as a u8. Fix it by explicitly casting to u8.

BUG=b:267963806
TEST=build ok

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4224897
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>

+++ u2fd/u2fhid.cc
@@ -308,7 +308,7 @@ int U2fHid::CmdPing(std::string* resp) {
 }
 
 int U2fHid::CmdLock(std::string* resp) {
-  int duration = transaction_->payload[0];
+  uint8_t duration = static_cast<uint8_t>(transaction_->payload[0]);
 
   VLOG(1) << "LOCK " << duration << "s CID:" << std::hex << transaction_->cid;
